+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 24.8.0
 - In the *AI settings* there now are buttons to test the connection to the AI services
   (for [#3062](https://github.com/pbek/QOwnNotes/issues/3062))
+- In the *AI settings* there now is a list of all AI backends from scripts and buttons
+  to test the connection to those AI services (for [#3064](https://github.com/pbek/QOwnNotes/issues/3064))
 
 ## 24.7.3
 - `gpt-4o-mini` was added to the list of available OpenAI models